Added systems

+++ src/ECS/ComponentManager.cpp
@@ -1,8 +1,8 @@
 #include"ComponentManager.h"
 
-std::vector<std::pair<unsigned int, std::shared_ptr<Component>>> ComponentManager::m_components = {};
+std::vector<std::pair<size_t, std::shared_ptr<Component>>> ComponentManager::m_components = {};
 
-std::vector<std::shared_ptr<Component>> ComponentManager::getAllObjectComponents(unsigned int idObjcet)
+std::vector<std::shared_ptr<Component>> ComponentManager::getAllObjectComponents(size_t idObjcet)
 {
 	std::vector<std::shared_ptr<Component>> components;
 	for (const auto& comp : m_components)
@@ -25,7 +25,7 @@ std::vector<std::shared_ptr<Component>> ComponentManager::getAllComponents()
 	return components;
 }
 
-void ComponentManager::addComponent(std::shared_ptr<Component> component, unsigned int idObject)
+void ComponentManager::addComponent(std::shared_ptr<Component> component, size_t idObject)
 {
 	m_components.push_back({ idObject,component });
 }
